+++ src/main/kotlin/com/github/etkachev/nxwebstorm/actionlisteners/DryRunAction.kt
@@ -16,14 +16,15 @@ class DryRunAction(
   private val id: String,
   private val formValues: FormValueMap,
   private val dialog: DialogWrapper,
-  private val type: SchematicTypeEnum
+  private val type: SchematicTypeEnum,
+  private val collectionPath: String
 ) : AbstractAction("Dry Run") {
   private var terminal = RunTerminalWindow(proj, "Dry Run")
   private var nxService = MyProjectService.getInstance(proj)
   override fun actionPerformed(e: ActionEvent?) {
     val values = formValues.formVal
     val projectType = nxService.nxProjectType
-    val command = getSchematicCommandFromValues(collection, id, values, projectType, true, type)
+    val command = getSchematicCommandFromValues(collection, id, values, projectType, true, type, collectionPath)
     dialog.close(DialogWrapper.CANCEL_EXIT_CODE)
     terminal.runAndShow(command)
   }